requirements.txt
# =============================================================================
# requirements.txt — speaker_focus project
# Python 3.10+ required
# Install: pip install -r requirements.txt
# =============================================================================

# Audio capture and playback
sounddevice>=0.4.6

# WebSocket client for Fixie.ai Ultravox
websockets>=12.0

# Signal processing — resampling
scipy>=1.11.0

# Numerical computing
numpy>=1.26.0

# Deep learning — MPS backend for Apple Silicon
torch>=2.2.0

# Speaker embedding model + inference wrapper
pyannote.audio>=3.1.0

# HuggingFace model loading
transformers>=4.40.0
huggingface-hub>=0.22.0

# Environment variable loading from .env files
python-dotenv>=1.0.0
